ajout de comfirmation pour supprimer

+++ views/utrip_manage.tpl
@@ -46,7 +46,7 @@
 						{if (isset($smarty.session.user.user_id) && ($smarty.session.user.user_role == "modo") || ($smarty.session.user.user_role == "admin"))}
 						<a class="btn btn-secondary" href="{$base_url}utrip/utrip?id={$objUtrip->getId()}" alt="Modérer l'article"><i class="fa fa-check-double"></i></a>
 						{/if}
-						<a class="btn btn-danger" href="{$base_url}utrip/delete?id={$objUtrip->getId()}" alt="Supprimer l'article"><i class="fa fa-trash"></i></a>
+						<a class="btn btn-danger" href="{$base_url}utrip/delete?id={$objUtrip->getId()}" onclick="return confirmDelete()" alt="Supprimer l'article"><i class="fa fa-trash"></i></a>
 					</td>
 				</tr>
 				{/foreach}
